+++ index.html
@@ -9,34 +9,18 @@
 <h1>All Movies</h1>
 
 <!-- Dropdown for Age Limit -->
-<label for="ageLimit">Select Age Limit:</label>
-<select id="ageLimit">
-    <option value="">All</option>
-    <option value="G">G</option>
-    <option value="PG">PG</option>
-    <option value="PG_13">PG-13</option>
-    <option value="R">R</option>
+<label for="category">Category:</label>
+<select id="category">
+    <option value="All">All</option>
+    <!-- Add other category options here -->
 </select>
 
-<!-- Dropdown for Category -->
-<label for="category">Select Category:</label>
-<select id="category">
-    <option value="">All</option>
-    <option value="ACTION">Action</option>
-    <option value="COMEDY">Comedy</option>
-    <option value="DRAMA">Drama</option>
-    <option value="ROMANCE">Romance</option>
-    <option value="HORROR">Horror</option>
-    <option value="SCIENCE_FICTION">Science Fiction</option>
-    <option value="FANTASY">Fantasy</option>
-    <option value="ANIMATION">Animation</option>
-    <option value="THRILLER">Thriller</option>
-    <option value="CRIME">Crime</option>
-    <option value="ADVENTURE">Adventure</option>
-    <option value="MYSTERY">Mystery</option>
-    <option value="DOCUMENTARY">Documentary</option>
-    <option value="FAMILY">Family</option>
+<label for="ageLimit">Age Limit:</label>
+<select id="ageLimit">
+    <option value="All">All</option>
+    <!-- Add other age limit options here -->
 </select>
+
 <button id="gotoEmployeeDashboard">Go to Employee Dashboard</button>
 <button id="login-button">Login</button>
